SITETB [IMPROVEMENT]: The visualization of cases per unit was modified when the user has the visualization of all administrative units. #43
GENERIC [IMPROVEMENT]: The system is now alerting when a batch is up to expire. The quantity of months that the system considers can be configured in the workspace setup. The (!) symbol is used to alert the batch that is up to expired and the (X) symbol is used to alert the batch that has already expired. #55

git-svn-id: https://www.etbmanager.org/svn/etbmanager/etbmanager/trunk@1903 d02e0694-cd7f-c14d-9e3c-2ffdef26b8fc